+++ example/data_source_virtual_environment_group.tf
@@ -2,10 +2,10 @@ data "proxmox_virtual_environment_group" "example" {
   group_id = "${proxmox_virtual_environment_group.example.id}"
 }
 
-output "data_proxmox_virtual_environment_group_example_comments" {
-  value = "${data.proxmox_virtual_environment_group.example.*.comment}"
+output "data_proxmox_virtual_environment_group_example_comment" {
+  value = "${data.proxmox_virtual_environment_group.example.comment}"
 }
 
 output "data_proxmox_virtual_environment_group_example_members" {
-  value = "${data.proxmox_virtual_environment_group.example.*.members}"
+  value = "${data.proxmox_virtual_environment_group.example.members}"
 }