+++ example/data_source_virtual_environment_users.tf
@@ -1,5 +1,5 @@
 data "proxmox_virtual_environment_users" "example" {
-  depends_on = ["proxmox_virtual_environment_role.example"]
+  depends_on = ["proxmox_virtual_environment_user.example"]
 }
 
 output "data_proxmox_virtual_environment_users_example_comments" {